+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Hiccup JVM compatibiltiy for fragments (see [#34](https://github.com/nextjournal/markdown/issues/34))
+* Frontmatter support (YAML, EDN and MultiMarkdown)
 
 ## 0.6.157